@@ -8,21 +8,24 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
-        "fideloper/proxy": "^4.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "5.8.*",
-        "laravel/tinker": "^1.0",
-        "maatwebsite/excel":"~2.1.0",
-        "ext-json": "*"
+        "php": "^8.1",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^10.10",
+        "laravel/tinker": "^2.8",
+        "laravel/ui": "^4.0",
+        "mews/captcha": "^3.0",
+        "morilog/jalali": "^3.1",
+        "rap2hpoutre/fast-excel": "^5.3.0",
+        "spatie/laravel-permission": "^6.3",
+        "tucker-eric/eloquentfilter": "^3.3"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "^1.0",
-        "filp/whoops": "^2.0",
-        "fzaninotto/faker": "^1.4",
-        "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^2.0",
-        "phpunit/phpunit": "^7.5"
+        "barryvdh/laravel-debugbar": "^3.11",
+        "fakerphp/faker": "^1.9.1",
+        "mockery/mockery": "^1.4.4",
+        "nunomaduro/collision": "^7.0",
+        "phpunit/phpunit": "^10.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -36,12 +39,10 @@
     },
     "autoload": {
         "psr-4": {
-            "studentPreRegisteration\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
