@@ -1,6 +1,6 @@
 <?php
 
-namespace studentPreRegisteration\Console;
+namespace App\Console;
 
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
