@@ -1,8 +1,8 @@
 <?php
 
-namespace studentPreRegisteration\Http\Controllers\Auth;
+namespace App\Http\Controllers\Auth;
 
-use studentPreRegisteration\Http\Controllers\Controller;
+use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\SendsPasswordResetEmails;
 
 class ForgotPasswordController extends Controller
@@ -14,7 +14,7 @@ class ForgotPasswordController extends Controller
     |
     | This controller is responsible for handling password reset emails and
     | includes a trait which assists in sending these notifications from
-    | your application to your users. Feel free to explore this trait.
+    | your application to your group_manager. Feel free to explore this trait.
     |
     */
 
