@@ -1,78 +1,78 @@
 <?php
 
-namespace studentPreRegisteration\Policies;
+namespace App\Policies;
 
-use studentPreRegisteration\User;
-use studentPreRegisteration\Course;
+use App\Models\User;
+use App\Models\Course;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class CoursePolicy
 {
     use HandlesAuthorization;
-    
+
     /**
      * Determine whether the user can view any courses.
      *
-     * @param  \studentPreRegisteration\User  $user
+     * @param  \App\Models\User  $user
      * @return mixed
      */
-    public function viewAll(User $user)
+    public function viewAny(User $user)
     {
-        return $user->role == 'groupManager';
+        return $user->hasRole('group_manager');
     }
 
     /**
      * Determine whether the user can view the course.
      *
-     * @param  \studentPreRegisteration\User  $user
-     * @param  \studentPreRegisteration\Course  $course
+     * @param  \App\Models\User  $user
+     * @param  \App\Models\Course  $course
      * @return mixed
      */
     public function view(User $user, Course $course)
     {
-        return $user->role == 'groupManager';
+        return $user->hasRole('group_manager');
     }
 
     /**
      * Determine whether the user can create courses.
      *
-     * @param  \studentPreRegisteration\User  $user
+     * @param  \App\Models\User  $user
      * @return mixed
      */
     public function create(User $user)
     {
-        return $user->role == 'groupManager';
+        return $user->hasRole('group_manager');
     }
 
     /**
      * Determine whether the user can update the course.
      *
-     * @param  \studentPreRegisteration\User  $user
-     * @param  \studentPreRegisteration\Course  $course
+     * @param  \App\Models\User  $user
+     * @param  \App\Models\Course  $course
      * @return mixed
      */
     public function update(User $user, Course $course)
     {
-        return $user->role == 'groupManager';
+        return $user->hasRole('group_manager');
     }
 
     /**
      * Determine whether the user can delete the course.
      *
-     * @param  \studentPreRegisteration\User  $user
-     * @param  \studentPreRegisteration\Course  $course
+     * @param  \App\Models\User  $user
+     * @param  \App\Models\Course  $course
      * @return mixed
      */
     public function delete(User $user, Course $course)
     {
-        return $user->role == 'groupManager';
+        return $user->hasRole('group_manager');
     }
 
     /**
      * Determine whether the user can restore the course.
      *
-     * @param  \studentPreRegisteration\User  $user
-     * @param  \studentPreRegisteration\Course  $course
+     * @param  \App\Models\User  $user
+     * @param  \App\Models\Course  $course
      * @return mixed
      */
     public function restore(User $user, Course $course)
@@ -83,8 +83,8 @@ class CoursePolicy
     /**
      * Determine whether the user can permanently delete the course.
      *
-     * @param  \studentPreRegisteration\User  $user
-     * @param  \studentPreRegisteration\Course  $course
+     * @param  \App\Models\User  $user
+     * @param  \App\Models\Course  $course
      * @return mixed
      */
     public function forceDelete(User $user, Course $course)
