@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'name' => 'سامانه انتخاب واحد',
+    'name' => env('APP_NAME','Laravel'),
 
     /*
     |--------------------------------------------------------------------------
@@ -80,7 +80,7 @@ return [
     |
     */
 
-    'locale' => 'en',
+    'locale' => 'fa',
 
     /*
     |--------------------------------------------------------------------------
@@ -165,15 +165,15 @@ return [
         /*
          * Package Service Providers...
          */
-        Maatwebsite\Excel\ExcelServiceProvider::class,
+        \Mews\Captcha\CaptchaServiceProvider::class,
         /*
          * Application Service Providers...
          */
-        studentPreRegisteration\Providers\AppServiceProvider::class,
-        studentPreRegisteration\Providers\AuthServiceProvider::class,
-        // studentPreRegisteration\Providers\BroadcastServiceProvider::class,
-        studentPreRegisteration\Providers\EventServiceProvider::class,
-        studentPreRegisteration\Providers\RouteServiceProvider::class,
+        App\Providers\AppServiceProvider::class,
+        App\Providers\AuthServiceProvider::class,
+        // App\Providers\BroadcastServiceProvider::class,
+        App\Providers\EventServiceProvider::class,
+        App\Providers\RouteServiceProvider::class,
 
     ],
 
@@ -225,8 +225,7 @@ return [
         'URL' => Illuminate\Support\Facades\URL::class,
         'Validator' => Illuminate\Support\Facades\Validator::class,
         'View' => Illuminate\Support\Facades\View::class,
-        'Excel' => Maatwebsite\Excel\Facades\Excel::class,
-
+        'Captcha' => \Mews\Captcha\Facades\Captcha::class
     ],
 
 ];
