@@ -37,7 +37,7 @@ return [
     |--------------------------------------------------------------------------
     |
     | This is the SMTP port used by your application to deliver e-mails to
-    | users of the application. Like the host we have set this value to
+    | group_manager of the application. Like the host we have set this value to
     | stay compatible with the Mailgun e-mail application by default.
     |
     */
