@@ -4,17 +4,18 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateTermsTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('terms', function (Blueprint $table) {
-            $table->integer('code')->unique();
+            $table->id();
+            $table->integer('code');
             $table->timestamps();
         });
     }
@@ -24,8 +25,8 @@ class CreateTermsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('terms');
     }
-}
+};
