@@ -31,7 +31,7 @@ require __DIR__.'/../vendor/autoload.php';
 | We need to illuminate PHP development, so let us turn on the lights.
 | This bootstraps the framework and gets it ready for use, then it
 | will load up this application so that we can run it and send
-| the responses back to the browser and delight our users.
+| the responses back to the browser and delight our group_manager.
 |
 */
 
