@@ -1,65 +1,45 @@
 @extends('layouts.app')
-
 @section('content')
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-md-8">
-            <div class="card">
-                <div class="card-header">{{ __('Reset Password') }}</div>
-
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="card col-5 p-0 border-primary">
+                <div class="card-header text-center text-white bg-primary">تغییر رمز عبور</div>
                 <div class="card-body">
-                    <form method="POST" action="{{ route('password.update') }}">
+                    <form action="{{route('auth.set-password')}}" method="post">
                         @csrf
-
-                        <input type="hidden" name="token" value="{{ $token }}">
-
-                        <div class="form-group row">
-                            <label for="email" class="col-md-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" value="{{ $email ?? old('email') }}" required autocomplete="email" autofocus>
-
-                                @error('email')
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $message }}</strong>
-                                    </span>
+                        <div class="my-4 row justify-content-center">
+                            <div class="form-group col-12 ">
+                                <label class="text-left">رمز عبور فعلی</label>
+                                <input type="password" class="form-control @error('current_password') is-invalid @enderror" name="current_password">
+                                @error('current_password')
+                                <span class="invalid-feedback" role="alert">
+                                    <strong>{{ $message }}</strong>
+                                </span>
+                                @enderror
+                            </div>
+                            <div class="form-group col-12">
+                                <label class="text-left">رمز عبور جدید</label>
+                                <input type="password" class="form-control @error('new_password') is-invalid @enderror" name="new_password">
+                                @error('new_password')
+                                <span class="invalid-feedback" role="alert">
+                                    <strong>{{ $message }}</strong>
+                                </span>
+                                @enderror
+                            </div>
+                            <div class="form-group col-12">
+                                <label class="text-left">تکرار رمز عبور جدید</label>
+                                <input type="password" class="form-control @error('confirm_password') is-invalid @enderror" name="confirm_password">
+                                @error('confirm_password')
+                                <span class="invalid-feedback" role="alert">
+                                    <strong>{{ $message }}</strong>
+                                </span>
                                 @enderror
                             </div>
                         </div>
-
-                        <div class="form-group row">
-                            <label for="password" class="col-md-4 col-form-label text-md-right">{{ __('Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" required autocomplete="new-password">
-
-                                @error('password')
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $message }}</strong>
-                                    </span>
-                                @enderror
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="password-confirm" class="col-md-4 col-form-label text-md-right">{{ __('Confirm Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required autocomplete="new-password">
-                            </div>
-                        </div>
-
-                        <div class="form-group row mb-0">
-                            <div class="col-md-6 offset-md-4">
-                                <button type="submit" class="btn btn-primary">
-                                    {{ __('Reset Password') }}
-                                </button>
-                            </div>
-                        </div>
+                        <input type="submit" class="btn btn-warning btn-block" value="ثبت" />
                     </form>
                 </div>
             </div>
         </div>
     </div>
-</div>
 @endsection
