@@ -7,15 +7,19 @@
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>{{ config('app.name', 'Laravel') }}</title>
+    <title>{{ config('app.name', 'Laravel') }} @hasSection('title') - @yield('title') @endif</title>
 
     <!-- Styles -->
     <link href="{{ asset('css/app.css') }}" rel="stylesheet">
 </head>
-<body class="rtl">
+<body @class(['rtl', 'full-bg' => request()->path() == 'login']) @if(request()->path() == 'login') style="background-image: url('{{asset('background.jpg')}}')" @endif>
     <div class="se-pre-con"></div>
     <div id="app">
-        <nav class="navbar navbar-expand-md navbar-light navbar-laravel">
+        <nav class="navbar navbar-expand-md navbar-light" style="background-color: @if(request()->path() == 'login') rgba(255, 255, 255, 0.88) @else #e3f2fd @endif">
+            @guest
+            <a class="navbar-brand" href="#">سامانه انتخاب واحد</a>
+            @endguest
+
             <div class="container">
                 <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="{{ __('Toggle navigation') }}">
                     <span class="navbar-toggler-icon"></span>
@@ -24,10 +28,9 @@
                 <div class="collapse navbar-collapse" id="navbarSupportedContent">
                     <!-- Left Side Of Navbar -->
                     <ul class="navbar-nav mr-auto">
-                        @auth
-                        @if(Auth::user()->role == 'groupManager')
-                            <li class="nav-item {{request()->routeIs('user.index')? 'active':''}}">
-                                <a class="nav-link" href="{{route('user.index')}}">
+                        @role('group_manager')
+                            <li class="nav-item {{request()->routeIs('admin.user.index')? 'active':''}}">
+                                <a class="nav-link" href="{{route('group-manager.user.index')}}">
                                     دانشجویان
                                 </a>
                             </li>
@@ -41,15 +44,16 @@
                                     تعریف انتخاب واحد
                                 </a>
                             </li>
-                        @elseif(Auth::user()->role == 'admin')
-                                <li class="nav-item {{request()->routeIs('user.index')? 'active':''}}">
-                                    <a class="nav-link" href="{{route('user.index')}}">
-                                        مدیران گروه
+                                <li class="nav-item {{request()->routeIs('report.index')? 'active':''}}">
+                                    <a class="nav-link" href="{{route('report.index')}}">
+                                        گزارش
                                     </a>
                                 </li>
-                                <li class="nav-item {{request()->routeIs('field.index')? 'active':''}}">
-                                    <a class="nav-link" href="{{route('field.index')}}">
-                                        رشته ها
+                        @endrole
+                        @role('admin')
+                                <li class="nav-item {{request()->routeIs('admin.user.index')? 'active':''}}">
+                                    <a class="nav-link" href="{{route('admin.user.index')}}">
+                                        مدیران گروه
                                     </a>
                                 </li>
                                 <li class="nav-item {{request()->routeIs('term.index')? 'active':''}}">
@@ -57,27 +61,34 @@
                                         سر ترم ها
                                     </a>
                                 </li>
-                        @endif
-                        @endauth
+                        @endrole
                     </ul>
 
                     <!-- Right Side Of Navbar -->
-                    <ul class="navbar-nav ml-auto">
-                        <!-- Authentication Links -->
-                        @auth
-                            <li class="nav-link"> {{Auth::user()->name . ' ' . Auth::user()->family}}</li>
-                            <li class="nav-item">
-                                <a class="nav-link" href=""
-                                   onclick="event.preventDefault();document.getElementById('logout-form').submit();">
-                                    خروج
-                                </a>
-                                <form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
-                                    @csrf
-                                </form>
-                            </li>
-                        @endauth
-                    </ul>
+
                 </div>
+                <ul class="navbar-nav ml-auto">
+                    <!-- Authentication Links -->
+                    @auth
+                        <li class="nav-link"> {{auth()->user()->name . ' ' . auth()->user()->family}}</li>
+                        <li class="nav-item">
+                            <button class="btn pointer"
+                                    onclick="event.preventDefault();document.getElementById('logout-form').submit();" data-toggle="tooltip" data-placement="top" title="خروج">
+                                <i class="fa fa-sign-out-alt fa-lg"></i>
+                            </button>
+                            <form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
+                                @csrf
+                            </form>
+                        </li>
+                        @role(['group_manager','admin'])
+                            <li class="nav-item">
+                                <a href="{{route('auth.reset-password')}}" data-toggle="tooltip" data-placement="top" title="تغییر رمز عبور">
+                                    <i class="fa fa-user mt-2 fa-lg"></i>
+                                </a>
+                            </li>
+                        @endrole
+                    @endauth
+                </ul>
             </div>
         </nav>
 
@@ -90,8 +101,20 @@
         $(window).on('load',function() {
             $(".se-pre-con").fadeOut("slow");
         });
+        $(function () {
+            $('[data-toggle="tooltip"]').tooltip()
+        })
+        $(document).ready(function () {
+            @if(session()->has('alert'))
+            Swal.fire({
+                title: "{{session('alert')['title']}}",
+                text: "{{session('alert')['message']}}",
+                type: "{{session('alert')['icon']}}",
+            });
+            @endif
+            $('*').persiaNumber();
+        })
     </script>
-    <script src='https://www.google.com/recaptcha/api.js'></script>
     @yield('script')
 </body>
 </html>
